requirements.txt
anyio==3.6.2
atomicwrites==1.4.1
attrs==22.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.62.0
google-search-results==2.4.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.2
orjson==3.8.9
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.0
rfc3986==1.5.0
rsa==4.5
six==1.15.0
sniffio==1.3.0
starlette==0.13.6
toml==0.10.2
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.25.10
uvicorn==0.12.3
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
